Home task  Rectangle is done ()

+++ dmitriy/src/main/hw_oop/Rectangle/Rectangle.java
@@ -23,11 +23,11 @@ public class Rectangle {
     }
 
     public double areaCalculator(double side1, double side2){
-        double area = (double) side1 * side2;
+        double area = side1 * side2;
         return area;
     }
     public double perimeterCalculator (double side1, double side2){
-        double perimeter = (double)(2 * (side1 + side2));
+        double perimeter = 2 * (side1 + side2);
         return perimeter;
     }
 }
